Fixed modularity and code to make it work
